move html entity replace to utils
